+++ src/ua/dp/daragan/MyChatServ.java
@@ -2,9 +2,7 @@
 //write "exit" for close
 package ua.dp.daragan;
 
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.net.ServerSocket;
 import java.net.Socket;
 import java.util.ArrayList;
@@ -12,61 +10,33 @@ import java.util.LinkedList;
 import java.util.Scanner;
 
 public class MyChatServ implements ClientsListner{
+    private static MyChatServ mcs = null;
     private ArrayList<Clients> clients;
     private LinkedList<String> allMsg;
     private int countOfMsgs = 0;
     private ServerSocket servSock = null;
     
-    Socket sock = null; //to do
 
-    public MyChatServ() {
-        clients = new ArrayList();
-    }
-    
-
-    public static void main(String[] args) {
-        MyChatServ mcs = new MyChatServ();
-        boolean exit = false;
-        
+    private MyChatServ() {
+        clients = new ArrayList<Clients>();
+        allMsg = new LinkedList<String>();
         try{
-            mcs.servSock = new ServerSocket(8080);
+            servSock = new ServerSocket(8080);
             System.out.println("Server started on port 8080!");
         } catch (IOException e){
             System.err.println(e.getStackTrace());
         }
-        
-        //--------------------------
-        one : while (true){
-           
-            try{
-                mcs.sock = mcs.servSock.accept();
-                Scanner sc = new Scanner( mcs.sock.getInputStream() );
-                
-                while(sc.hasNextLine()){
-                    String str1 = sc.nextLine();
-                    if(str1.equalsIgnoreCase("exit") ) break one;
-                    System.out.println(str1);   
-                }
-                mcs.sock.close();
-
-            } catch (IOException e){
-                System.err.println(e.getStackTrace());
-            } 
+    }
+    
+    public static MyChatServ getInstance(){
+        if(mcs == null) {
+            mcs = new MyChatServ();
         }
-        //--------------------------
-        
-        
-        Client a,b; //tests
-        a = new Client(mcs);
-        b = new Client(mcs);
-        mcs.sendToAll();
-        
-        try{
-            mcs.servSock.close();
-            System.out.println("Server stopped!");
-        }catch (IOException ioe) {
-            System.err.println(ioe.getStackTrace());
-        }
+        return mcs;
+    }
+    
+    public ServerSocket getServSock (){
+        return servSock;
     }
 
     @Override
